requirements.txt
asgiref==3.8.1
dj-database-url==3.0.0
Django==5.2.3
django-environ==0.12.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
sqlparse==0.5.3
typing_extensions==4.14.0
